+++ API/src/main/java/com/svalero/API/service/GardenerService.java
@@ -2,6 +2,7 @@ package com.svalero.API.service;
 
 import com.svalero.API.domain.Gardener;
 import com.svalero.API.domain.dto.GardenerDto;
+import com.svalero.API.exceptions.GardenerNotFoundException;
 import com.svalero.API.repository.GardenerRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,8 +20,11 @@ public class GardenerService {
     public List<Gardener> getAll(){
         return this.gardenerRepository.findAll();
     }
-    public Gardener getById(long id){
-        return this.gardenerRepository.findById(id);
+    public Gardener getById(long id) throws GardenerNotFoundException {
+        Gardener gardener = this.gardenerRepository.findById(id);
+        if(gardener == null)
+            throw new GardenerNotFoundException("Gardener with ID " + id + " not found");
+        return gardener;
     }
 
     public Gardener add(GardenerDto newGardener) {
@@ -28,14 +32,18 @@ public class GardenerService {
         return this.gardenerRepository.save(gardener);
     }
 
-    public int remove(long id){
-        //this.gardenerRepository.findById(id);
+    public int remove(long id) throws GardenerNotFoundException {
+        Gardener gardener = this.gardenerRepository.findById(id);
+        if(gardener == null)
+            throw new GardenerNotFoundException("Gardener with ID " + id + " not found");
         this.gardenerRepository.deleteById(id);
         return 1;
     }
 
-    public Gardener modify(long id, GardenerDto updatedGardener){
+    public Gardener modify(long id, GardenerDto updatedGardener) throws GardenerNotFoundException {
         Gardener gardener = this.gardenerRepository.findById(id);
+        if(gardener == null)
+            throw new GardenerNotFoundException("Gardener with ID " + id + " not found");
         modelMapper.map(updatedGardener, gardener);
         return this.gardenerRepository.save(gardener);
     }
